Single client for Kibana
