+++ internal/kibana/dashboards.go
@@ -2,13 +2,12 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package dashboards
+package kibana
 
 import (
 	"encoding/json"
 	"io/ioutil"
 	"net/http"
-	"os"
 	"strings"
 
 	"github.com/pkg/errors"
@@ -16,35 +15,10 @@ import (
 	"github.com/elastic/elastic-package/internal/common"
 	"github.com/elastic/elastic-package/internal/logger"
 	"github.com/elastic/elastic-package/internal/multierror"
-	"github.com/elastic/elastic-package/internal/stack"
 )
 
 type exportedType struct {
 	Objects []common.MapStr `json:"objects"`
-}
-
-// Client is responsible for exporting dashboards from Kibana.
-type Client struct {
-	host     string
-	username string
-	password string
-}
-
-// NewClient creates a new instance of the client.
-func NewClient() (*Client, error) {
-	host := os.Getenv(stack.KibanaHostEnv)
-	if host == "" {
-		return nil, stack.UndefinedEnvError(stack.KibanaHostEnv)
-	}
-
-	username := os.Getenv(stack.ElasticsearchUsernameEnv)
-	password := os.Getenv(stack.ElasticsearchPasswordEnv)
-
-	return &Client{
-		host:     host,
-		username: username,
-		password: password,
-	}, nil
 }
 
 // Export method exports selected dashboards using the Kibana Export API.
